Refactor avr-hal-generic to move mega-specific code into atmega-hal
